Added create and delete guild function

+++ src/net/command/chat/CommandSet.java
@@ -12,7 +12,7 @@ public class CommandSet extends Command {
 	@Override
 	public void read(Player player) {
 		Connection connection = player.getConnection();
-		byte packetID = connection.readByte();
+		short packetID = connection.readShort();
 		int id = connection.readInt();
 		int value = connection.readInt();
 		Player member = id == player.getUnitID() ? player : Server.getInGameCharacter(id);
